pyproject.toml
[project]
name = "krapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "ishida", email = "ishida@laboro.ai" }
]
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.2.3",
    "pyyaml>=6.0.2",
    "sqlalchemy>=2.0.40",
    "streamlit>=1.44.1",
    "typer>=0.15.2,<0.16.0",
]

[project.scripts]
krapp = "krapp.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pytest>=8.3.5",
]
